Rename expr.rs in prime.rs
